move remaining components in to components directory
